+++ apps/web/src/rpc/wsRpcClient.ts
@@ -187,12 +187,6 @@ export interface WsRpcClient {
     readonly deleteTask: RpcUnaryMethod<typeof ORCHESTRATOR_WS_METHODS.deleteTask>;
     readonly clearPmChat: RpcUnaryMethod<typeof ORCHESTRATOR_WS_METHODS.clearPmChat>;
     readonly requestPmHandoff: RpcUnaryMethod<typeof ORCHESTRATOR_WS_METHODS.requestPmHandoff>;
-    readonly getProjectContextOnboarding: RpcUnaryMethod<
-      typeof ORCHESTRATOR_WS_METHODS.getProjectContextOnboarding
-    >;
-    readonly dismissProjectContextOnboarding: RpcUnaryMethod<
-      typeof ORCHESTRATOR_WS_METHODS.dismissProjectContextOnboarding
-    >;
     readonly requestProjectContextRun: RpcUnaryMethod<
       typeof ORCHESTRATOR_WS_METHODS.requestProjectContextRun
     >;
@@ -204,15 +198,6 @@ export interface WsRpcClient {
     >;
     readonly getProjectContextRunReview: RpcUnaryMethod<
       typeof ORCHESTRATOR_WS_METHODS.getProjectContextRunReview
-    >;
-    readonly reviseProjectContextRun: RpcUnaryMethod<
-      typeof ORCHESTRATOR_WS_METHODS.reviseProjectContextRun
-    >;
-    readonly commitProjectContextRun: RpcUnaryMethod<
-      typeof ORCHESTRATOR_WS_METHODS.commitProjectContextRun
-    >;
-    readonly discardProjectContextRun: RpcUnaryMethod<
-      typeof ORCHESTRATOR_WS_METHODS.discardProjectContextRun
     >;
     readonly getLaunchCapabilities: RpcUnaryNoArgMethod<
       typeof ORCHESTRATOR_WS_METHODS.getLaunchCapabilities
@@ -445,14 +430,6 @@ export function createWsRpcClient(transport: WsTransport): WsRpcClient {
         transport.request((client) => client[ORCHESTRATOR_WS_METHODS.clearPmChat](input)),
       requestPmHandoff: (input) =>
         transport.request((client) => client[ORCHESTRATOR_WS_METHODS.requestPmHandoff](input)),
-      getProjectContextOnboarding: (input) =>
-        transport.request((client) =>
-          client[ORCHESTRATOR_WS_METHODS.getProjectContextOnboarding](input),
-        ),
-      dismissProjectContextOnboarding: (input) =>
-        transport.request((client) =>
-          client[ORCHESTRATOR_WS_METHODS.dismissProjectContextOnboarding](input),
-        ),
       requestProjectContextRun: (input) =>
         transport.request((client) =>
           client[ORCHESTRATOR_WS_METHODS.requestProjectContextRun](input),
@@ -468,18 +445,6 @@ export function createWsRpcClient(transport: WsTransport): WsRpcClient {
       getProjectContextRunReview: (input) =>
         transport.request((client) =>
           client[ORCHESTRATOR_WS_METHODS.getProjectContextRunReview](input),
-        ),
-      reviseProjectContextRun: (input) =>
-        transport.request((client) =>
-          client[ORCHESTRATOR_WS_METHODS.reviseProjectContextRun](input),
-        ),
-      commitProjectContextRun: (input) =>
-        transport.request((client) =>
-          client[ORCHESTRATOR_WS_METHODS.commitProjectContextRun](input),
-        ),
-      discardProjectContextRun: (input) =>
-        transport.request((client) =>
-          client[ORCHESTRATOR_WS_METHODS.discardProjectContextRun](input),
         ),
       getLaunchCapabilities: () =>
         transport.request((client) => client[ORCHESTRATOR_WS_METHODS.getLaunchCapabilities]({})),